Volumetric is working, though changes that were made to have it reference multiple file locations is not correct.
All files related to a top-level object should be stored in the storage provider under that
entity. e.g., all files for a volumetric (mp4, drcs, manifest) should be in /volumetric/<volumetric_id>
The mp4, for instance, should not be in /video/<id>; only things that are videos should be in /video

Do not use hashes in file keys. Use the ID of the top-level object instead. The hash in the database
entry for the static resource is enough.

Started converting images. Not done.